requirements.txt
# AI GURU - Personalized RAG Research Assistant
# Core dependencies

# Web UI
streamlit>=1.30.0

# LLM Provider
anthropic>=0.18.0

# Environment management
python-dotenv>=1.0.0

# Vector Database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.3.0

# Document Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Text Processing
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Utilities
tiktoken>=0.5.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
